Добавлена возможность получать доступ к любым полям любого класса в ast, а также узнавать тип этого класса

+++ 5lab/compilers-lab5-src/examples/test.cl
@@ -4,7 +4,7 @@ class B inherits A {};
 class B {};
 
 class Main {
-    aboba : Int <- 10;
+    aboba : Float <- 10;
     main(): Int {
         10
     };
